Allow specifying the underlying list-box% selection type
The underlying list-box% control allows specifying whether to allow selecting
a single item or multiple items.  Expose this argument as the `selection-type`
init field of the qresults-list% object.

Also added a `get-selected-row-indexes` method to retrieve the indexes of all
selected rows (or an empty list if no rows are selected).